data_pipeline/requirements.txt
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
googlemaps==4.6.0
pandas==2.2.1
geopandas==0.14.3
tqdm==4.66.2